split http endpoint tests into a separate file
